requirements.txt
flask==3.1.1
flask-cors
gunicorn
python-dotenv==1.1.0
sentence-transformers==4.1.0
pypdf==5.6.1
langchain==0.3.26
langchain-community==0.3.26
langchain-openai==0.3.24
langchain-pinecone==0.2.8
langchain-groq
langchain-huggingface
pinecone
torch
transformers
huggingface-hub